Quite some refactoring ongoing
